+++ CHANGELOG.md
@@ -3,10 +3,22 @@
 All notable changes to this module are documented in this file.
 cryptox module adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.2.0] - 2014-02-15
+## [0.2.1] - Unreleased
+
 ### Fixed
-- Fix bitstamp.getFee() was not returning error when API keys invalid
-- 
+- Fix JSON validation schema for `getBalance()` result
+- Fix: `bitfinex.getTicker()` now returns the result with correct JSON schema
+
+### Added
+
+### Changed
+
+
+## [0.2.0] - 2014-02-15
+
+### Fixed
+- Fix: `bitstamp.getFee()` now returns an error when invalid API keys 
+
 ### Added
 - Add parameter `account_id` to `getBalance()` response
 - Implement `getBalance()` method for Bitstamp